Separated Login and Search component
