+++ my-app/src/components/SearchBar.css
@@ -16,11 +16,6 @@
   padding-right: 10px;
 }
 
-.login-container {
-  display: flex;
-  align-items: center;
-}
-
 .form-field {
   flex: 1;
   display: flex;
@@ -41,108 +36,4 @@
   background-color: #eeeeee;
   padding-left: 20px;
   border-radius: 30px;
-}
-
-.login-popup {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  height: 220px;
-  width: 280px;
-  transform: translate(-50%, -50%);
-  background-color: #fff;
-  padding: 20px;
-  box-shadow: 4px 4px 10px rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-  z-index: 999;
-}
-
-.login-form {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.login-form input {
-  width: 100%;
-  height: 40px;
-  margin-bottom: 10px;
-  padding: 10px 20px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  outline: none;
-}
-
-.login-form button {
-  width: 100%;
-  height: 40px;
-  background-color: var(--theme);
-  color: #fff;
-  font-size: 16px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  outline: none;
-}
-
-.login-form button:hover {
-  background-color: var(--selected);
-  transition: background-color 0.3s ease;
-}
-
-.login-button {
-  width: 120px;
-  height: 40px;
-  background-color: var(--theme);
-  color: #fff;
-  font-size: 16px;
-  border: none;
-  border-radius: 20px;
-  cursor: pointer;
-  outline: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  transition: background-color 0.3s ease;
-  margin-left: 5px;
-}
-
-.login-button:hover {
-  background-color: var(--selected);
-}
-
-.login-button:focus {
-  outline: none;
-}
-
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(213, 213, 213, 0.238);
-  backdrop-filter: blur(5px);
-  z-index: 999;
-}
-
-.close-button {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  background: none;
-  border: none;
-  font-size: 1.5rem;
-  cursor: pointer;
-  color: #888;
-}
-
-.close-button:hover {
-  color: var(--selected);
-  transition: color 0.3s ease;
 }
